requirements.txt
uvicorn
asyncpg
databases
python-dotenv
jinja2
fastapi>=0.104.0
uvicorn>=0.23.2
langchain>=0.0.335
langchain-community>=0.0.13
langchain-openai>=0.0.5
langgraph>=0.0.20
pydantic>=2.4.2
python-dotenv>=1.0.0
twilio>=8.10.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
pyodbc>=4.0.39
urllib3>=2.0.0 
